Create R2DBC module with application and tests
This commit introduces an entire R2DBC module which includes the basic structure for a Spring Boot application, a "Student" entity with an R2DBC repository. Tests for the application, repository and an entity-dto mapper are also included. Besides, it contains a schema and data SQL files for initializing the database, YAML configurations for the application and its tests, and Gradle build specifications. The lombok.config and .gitignore files are adjusted for the new module as well. Finally, the settings.gradle.kts is modified to include the new R2DBC module under the root project.