Add private constructor to StudentMapper utility class
A private constructor is added to StudentMapper to prevent instantiation. This is a utility class, thus it's not designed to be instantiated, because it consists only of static methods. Adding a private constructor enhances the class by ensuring the Java garbage collector won't have to deal with unnecessary object allocation.